Refactor utility tests.
