+++ tests/services/api_tests/model_test.py
@@ -16,7 +16,7 @@
 import unittest
 import json
 from tests.services.api_tests.api_tester import ModelTestRunner
-from tests.services.utils.db import create_test_engine
+from tests.services.util.db import create_test_engine
 from tests.unittest_utils import ForsetiTestCase
 from google.cloud.forseti.services.dao import ModelManager
 from google.cloud.forseti.services.explain.service import GrpcExplainerFactory
