+++ tests/services/inventory/storage_test.py
@@ -15,7 +15,7 @@
 
 import unittest
 from tests.unittest_utils import ForsetiTestCase
-from tests.services.utils.db import create_test_engine
+from tests.services.util.db import create_test_engine
 from google.cloud.forseti.services import db
 from google.cloud.forseti.services.inventory.base.resources import Resource
 from google.cloud.forseti.services.inventory.storage import initialize
